Here i'm adding some files, basically a base arquitecture, i hope you like it
