+++ templates/homepage.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<title>FibonacciApp</title>
+</head>
+<body>
+	<h1>Fibonacci</h1>
+</body>
+</html>
